Simple system of task scripts to replace Rakefile
We still have a dependency on a ruby gem, pfm, but that can be
eliminated once we have a node.js port of Panda Flavored Markdown.

Basic usage of tasks:

    coffee task/build # runs all scripts in this dir
    coffee task/build/docs.coffee

Tab-completion is your friend here.

Some of these are wired up to the npm lifecycle scripts (as declared
in package.json)